Moved tests to approriate files
